Move group selection to components
